Remove ServiceGroup - Refactor Header out from Frame
