--- .favorites.json
@@ -1,10 +1,0 @@
-[
-  {
-    "type": "Directory",
-    "name": "/Users/dongkyun/Documents/Projects/Whatssub Admin/dooboo-admin/assets/icons",
-    "parent_id": null,
-    "workspaceRoot": "/Users/dongkyun/Documents/Projects/Whatssub Admin/dooboo-admin",
-    "workspacePath": "assets/icons",
-    "id": "qhOf3d0Ip7zH21ym"
-  }
-]
