--- src/components/screen/ServiceGroup/styles.css
@@ -1,16 +1,0 @@
-.pg-ul {
-  display: flex;
-  flex-direction: row;
-  list-style-type: none;
-}
-
-.pg-item {
-  color: #8b8c8c;
-  text-decoration: none;
-  padding: 0.5em;
-  font-size: 0.8em;
-}
-
-.pg-active-item {
-  color: #407fff;
-}
