+++ src/components/ui/Header/Header.tsx
@@ -5,25 +5,9 @@ import { NavLink } from 'react-router-dom';
 import { ROUTE } from '../../../utils/Constants';
 import { RouteContext } from '../../../providers/RouteProvider';
 import Title from '../../shared/Title';
-import { WhenNarrow } from '../../../utils/Styles';
 import styled from 'styled-components';
 
-const fontSize = `20px`;
-
-const Container = styled.div`
-  width: 100vw;
-  height: 100%;
-  padding: 3rem;
-  box-sizing: border-box;
-  ${WhenNarrow} {
-    padding: 1rem;
-  }
-  background: ${(props) => props.theme.background};
-  color: ${(props) => props.theme.fontColor};
-  font-size: ${fontSize};
-`;
-
-const Header = styled.div`
+const NavList = styled.div`
   margin-top: 0.5rem;
   margin-bottom: 1.5rem;
   /* background-color: red; */
@@ -42,11 +26,11 @@ const activeStyle = {
   color: '#407fff',
   borderBottom: '1px solid #407fff',
 } as React.CSSProperties;
-const Frame: React.FC = (props) => {
+const Header: React.FC = (props) => {
   const { route } = useContext(RouteContext);
   const { isSignedIn } = useContext(AuthContext);
   return (
-    <Container>
+    <>
       <Title
         onLogoClick={() => {
           route.history.push(ROUTE.main);
@@ -56,14 +40,11 @@ const Frame: React.FC = (props) => {
         }
         larger
       />
-      <Header>
+      <NavList>
         {isSignedIn ? (
           <>
             <CNavLink to={ROUTE.main} activeStyle={activeStyle}>
               서비스
-            </CNavLink>
-            <CNavLink to={ROUTE.serviceGroup} activeStyle={activeStyle}>
-              서비스그룹
             </CNavLink>
             <CNavLink to={ROUTE.category} activeStyle={activeStyle}>
               카테고리
@@ -74,10 +55,8 @@ const Frame: React.FC = (props) => {
             로그인
           </CNavLink>
         )}
-      </Header>
-
-      {props.children}
-    </Container>
+      </NavList>
+    </>
   );
 };
-export { Frame };
+export default Header;
