+++ src/contexts/index.ts
@@ -1,9 +1,6 @@
 import * as React from 'react';
-import { IUser, ILocale } from '../types';
 
 interface IAppContext {
-  user?: IUser;
-  locale?: ILocale;
   state?: any;
   dispatch?: any;
 }
@@ -14,6 +11,8 @@ interface IServiceModalContext {
 
 const AppContext = React.createContext<IAppContext | null>(null);
 
-const ServiceModalContext = React.createContext<IServiceModalContext | null>(null);
+const ServiceModalContext = React.createContext<IServiceModalContext | null>(
+  null,
+);
 
 export { AppContext, ServiceModalContext };
